+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.98.0 (Unreleased)
 FEATURES:
 * mongodb: support `performance_diagnostics` in `yandex_mdb_mongodb_cluster`
+* serverless: support `stream_names` attribute in logging triggers
 
 FEATURES:
 * k8s: added `gpu_environment` property to `gpu_settings` attribute of `node_group` resource and data source